# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cerita",
        "description": "API Dokumentasi dengan Sanctum Bearer Token",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local API Server"
        },
        {
            "url": "https://capstone.andreasadi.my.id",
            "description": "Deployed API Server"
        }
    ],
    "paths": {
...
        "/api/analyze-image": {
            "post": {
                "tags": [
                    "Analisis Gambar"
                ],
                "summary": "Analisis gambar menggunakan model AI",
                "description": "Mengirim gambar (upload file atau base64) untuk dianalisis oleh model AI dan mengembalikan hasil klasifikasi.",
                "operationId": "analyzeImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "File gambar (jpg, jpeg, png)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "camera_image"
                                ],
                                "properties": {
                                    "camera_image": {
                                        "description": "Base64 dari gambar hasil kamera",
                                        "type": "string",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil menganalisis gambar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil menganalisis gambar"
                                        },
                                        "uploaded_image": {
                                            "type": "string",
                                            "example": "http://localhost:8000/uploads/xxx.png"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (gambar tidak valid)"
                    },
                    "422": {
                        "description": "Validasi gagal (gambar tidak sesuai)"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ceritas": {
            "get": {
                "tags": [
                    "Cerita"
                ],
                "summary": "Get all cerita with images",
                "operationId": "ee5c0842fd7ceb2cfe64a9f24fad1fe0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data cerita berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cerita"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized, please login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ceritas/{id}": {
            "get": {
                "tags": [
                    "Cerita"
                ],
                "summary": "Get single cerita by id",
                "operationId": "e2ce6a7995a0594b23879e0464a3444e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID cerita",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data cerita berhasil diambil"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cerita"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data cerita tidak ditemukan"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized, please login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Image": {
                "title": "Image",
                "description": "Gambar terkait cerita",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/images/image1.jpg"
                    },
                    "cerita_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-07T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-07T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Cerita": {
                "title": "Cerita",
                "description": "Data cerita lengkap dengan images",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Judul Cerita"
                    },
                    "content": {
                        "type": "string",
                        "example": "Isi cerita yang lengkap..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-07T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-07T10:00:00Z"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Image"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}